# src/app/api/apyhub/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://reformify.dev",
      "description": "Production server"
    }
  ],
  "info": {
    "title": "Reformify API",
    "description": "API for managing forms and submissions",
    "version": "1.0.0"
  },
  "paths": {
    "/api/apyhub/create_form": {
      "post": {
        "summary": "Create a new form",
        "security": [
          {
            "customerId": [],
            "apyAuthorization": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["formKey", "title", "elements"],
                "properties": {
                  "formKey": {
                    "type": "string",
                    "description": "Identifier for the form that is unique to the user's other forms"
                  },
                  "title": {
                    "type": "string",
                    "description": "Form title"
                  },
                  "description": {
                    "type": "string",
                    "description": "Form description"
                  },
                  "elements": {
                    "type": "array",
                    "description": "Form elements/fields",
                    "items": {
                      "type": "object",
                      "required": ["type", "name"],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["text", "number", "email", "date", "boolean"]
                        },
                        "name": {
                          "type": "string"
                        },
                        "required": {
                          "type": "boolean",
                          "default": false
                        }
                      }
                    }
                  },
                  "emailSettings": {
                    "type": "object",
                    "properties": {
                      "requireEmail": {
                        "type": "boolean",
                        "description": "Whether email is required for submission",
                        "default": false
                      },
                      "confirmationEmail": {
                        "type": "string",
                        "description": "Custom confirmation email text"
                      },
                      "notifyOnEntry": {
                        "type": "boolean",
                        "description": "Whether to notify admin via email on new submissions",
                        "default": false
                      },
                      "sendConfirmation": {
                        "type": "boolean",
                        "description": "Whether to send confirmation emails to submitters - requires requireEmail to be true",
                        "default": false
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Form created successfully",
            "headers": {
              "x-apy-atoms": {
                "schema": {
                  "type": "string"
                },
                "description": "Number of atoms consumed by this request"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "acknowledged": {
                      "type": "boolean"
                    },
                    "insertedId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Form key already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/apyhub/submission": {
      "post": {
        "summary": "Submit a form entry",
        "security": [
          {
            "customerId": [],
            "apyAuthorization": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["formKey", "entry"],
                "properties": {
                  "formKey": {
                    "type": "string"
                  },
                  "entry": {
                    "type": "object",
                    "description": "Form submission data - must match the form's defined elements",
                    "additionalProperties": {
                      "oneOf": [
                        {
                          "type": "string",
                          "description": "For text, email, and date fields"
                        },
                        {
                          "type": "number",
                          "description": "For number fields"
                        },
                        {
                          "type": "boolean",
                          "description": "For boolean fields"
                        }
                      ]
                    },
                    "example": {
                      "name": "John Doe",
                      "email": "john@example.com",
                      "age": 25,
                      "birthDate": "1998-01-01",
                      "subscribe": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry submitted successfully",
            "headers": {
              "x-apy-atoms": {
                "schema": {
                  "type": "string"
                },
                "description": "Number of atoms consumed by this request"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "entryId": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation error"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/apyhub/get_submissions": {
      "post": {
        "summary": "Get form submissions",
        "security": [
          {
            "customerId": [],
            "apyAuthorization": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["formKey"],
                "properties": {
                  "formKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of form submissions",
            "headers": {
              "x-apy-atoms": {
                "schema": {
                  "type": "string"
                },
                "description": "Number of atoms consumed by this request"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "formId": {
                        "type": "string"
                      },
                      "formName": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "entry": {
                        "type": "object"
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Form not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "customerId": {
        "type": "apiKey",
        "name": "x-customer-id",
        "in": "header",
        "description": "Customer ID provided to authenticate requests. Must be included in the x-customer-id header."
      },
      "apyAuthorization": {
        "type": "apiKey",
        "name": "x-apy-authorization",
        "in": "header",
        "description": "Shared secret for APY authorization. Must match the server's configured secret."
      }
    }
  }
}